% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DualScaling.R
\name{DualScaling}
\alias{DualScaling}
\title{Dual Sacling with EigenDecomposition}
\usage{
DualScaling(dat)
}
\arguments{
\item{dat}{A matrix or data frame}
}
\description{
Dual Sacling with EigenDecomposition
}
