% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DualScalingMA.R
\name{DualScalingMA}
\alias{DualScalingMA}
\title{Dual Sacling with Mutual Averages}
\usage{
DualScalingMA(dat, eps = 1e-10, iter.max = 1000)
}
\arguments{
\item{dat}{A matrix or data frame}

\item{eps}{A criteria whether it reached conversion. Default value is 10e-5}

\item{iter.max}{A muximum numbers of iterations}
}
\description{
Dual Sacling with Mutual Averages
}
