% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DualScalingOrdinal.R
\name{DualScalingOrdinal}
\alias{DualScalingOrdinal}
\title{Dual Sacling for Oridinal Data}
\usage{
DualScalingOrdinal(dat)
}
\arguments{
\item{dat}{A matrix or data frame}
}
\description{
Dual Sacling for Oridinal Data
}
